Containerize the codecov uploader
Since there isn't an existing container for the uploader, we'll make
our own. The version of the uploader is parameterizable, and
incorporated into the tag of the container, making it easy to select a
specific version to use, should that be desired.

The uploader binary itself is cryptographically verified using
Codecov's PGP keys. The resulting container is minimal, based on the
`distroless/cc` image from Google.

The plugin is also parameterizable, allowing consumers to either use
the latest available container (the default), or to pin to a specific
version (or use another container altogether).

Signed-off-by: Christopher Maier <chris@graplsecurity.com>