Add a `fail_job_on_error` configuration parameter
Before adding the `--rootDir` option to our invocations in #26, we
were silently failing to upload code coverage statistics to Codecov
for two weeks.

By default, `codecov` always returns an exit code of `0`, whether it
succeeded or failed, specifically to _not_ impact CI jobs. However,
this also can make it difficult to see when reporting of statistics
has failed!

The `codecov` uploader _does_ have a `--nonZero` flag, however, which
will enable failures to return a non-zero exit code, should you
desire.

Here, we add a new `fail_job_on_error` parameter which governs whether
or not a failure of this plugin will fail the Buildkite job it is part
of. If the value is `true`, we add the `--nonZero` option; if not, we
don't.

We will default to `true` for the time being.

If this turns out to be problematic, it can be easily counteracted
globally in a given pipeline by adding the following to the top
of the pipeline definition:

```yaml
env:
  BUILDKITE_PLUGIN_CODECOV_FAIL_JOB_ON_ERROR: false
```

A potential improvement to be made later could be to add retries to
lessen the impact of any transient network issues.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>