Update base build image to `ubuntu:22.04`
Here, we update our base image to `ubuntu:22.04` ("Jammy Jellyfish"
LTS). The previous Ubuntu version we were using, 21.04 ("Hirsute
Hippo") was end-of-lifed on January 20, 2022, and the apt repository
links for it apparently just went dead over the weekend.

The final image we ship continues to be based on a `distroless` image,
but we still have to download and verify the `codecov` binary, so this
ensures our "build environment" continues to work.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>